feature/get emotion from Page

+++ src/main/java/cau/capstone/backend/global/aiserver/EmotionDto.java
@@ -13,6 +13,6 @@ public class EmotionDto {
 
 
     String emotion;
-    double value;
+    int value;
 
 }
